Improve error handling in multiple instances.
The script had some very old code concept to cat the target response file on error in various instances. But that concept is now much behind. Plus, after we revetered curl option --fail-with-body to --fail only, old buffers would be catted on error, making debugging confusing (for me). So now we removed all cat commands that would execute on error to minimise dumping old contents to user on stderr.

Also, ImageMagick 7 deprecated the convert command, so we check for it in the hope to minimise the deprecation notice from IM. That should still work on IM 6 and above normally.

---