New Bash and Zsh completion scripts! Minor fixes and improvements.
Completion scripts for Bash and Zsh are now available for command line completion of options and arguments! See updated documentation!

Add basic multi modality for text models. This means that from now on, if the last argument is a text or pdf file, or a recognisable url, the script will dump and add it to the tail of the current user prompt before making a call.

I find this to be very useful myself!

Improve globs of functions to try and detect file types.

The redo user feedback command in the confirmation prompt will clear the input, as it did a long time ago (this behaviour had been lost but is definitely interesting to bring back..).

Add chat command `!pdf` or `/pdf` to dump pdf text. There are some optional dependencies for this (check the updated docs). It will also read the pdf as text when setting it as the last command line option (before it would just add the dump of plain text files)

Command "/cat" will redirect to "/pdf" in case it detects a pdf input, and will also redirect to "/url" if that is a url.

Fix to print up to 1 million tokens of previous session from the history, this affects mostly option -HH.

Option -vv in conjunction with -HH to print the history will without asking about and confirming which session the user wants to print!

Add option -EE to better work with the completion functions.

Update man, docs, and readme.md.

---