Fixes editor mode option -x when multiple instances are running, fix recoding text/pdf/url dumps to history.
Fix setting what would be recorded to the tsv history file after the new features of appending text, pdf and url dumps after the last argument of the user prompt when it is text, pdf or url. The dump would be submitted once however what would be recorded to the history file was not being updated in all cases, so the following turns of the conversation would not have the content of the dump anymore, of the filename for example.

Note the script processes these and hard record the dump to the history file, as doing otherwise would conflict with the multimodel way of processing files (image files  and image urls).

Use double newlines to separate the filename / url (last argument of user prompt) to the file / url dump.

Update Google AI defaults model to gemini-1.5-flash-latest (gemini-1.5-pro is better but for me it is rate limited to the point I only use it when I need a better response).

Improve trim_leadf() and trim_trailf() default index from 160 to 320 chars as I had some problems with the new features because this value was too small to properly clean some inputs. 320 chars should add reliability to the functions without compromising speed.

Move set_filetxtf() in the main editor wrapper function to set the filename of the file for the text editor before we would inject the pre prompt...

---