Some good fixes from the text editor interface and some general fixes
Since we removed the curl --fail-with-body option, when curl got a server error,
a previous json response from a valid requests would be read as though there was no error in curl.

We now clear curl output file before the request, so that if it fails, the output file will always be empty.

Also, curl exit code was not being propagated to the main functions any more (because we started
relying on --fail-with-body some time ago but now that is not used anymore), now fixed!

Add new option for retrying whisper request on chat mode in the user conformation prompt.
Before, i had already implemented the chat command "/request" which would resend the
last recorded audio to whisper to serve as input. But it is easier to retry a new whisper request
with the already recorded audio now  in the confirmation question the script makes to the user
(only on verbose modes, option -v must not be set).

lastjsonf() function (option -Z) will either print the actual last response file or if empty, it will print the
older response instead and exit with 1.

We will print to stderr the name of the cli browser or curl when retrieving url dumps as this may take a second or two.

Fix and Improve the preview function. It will recognise when the last argument of user input is a URL and a directory, meaning it will only be activated in the proper cases.

So to have a URL dumped, do as in the following prompt:

   Please make an abstract of the contents. https://mountaineerbr.github.io/blog/2/

However, to have the preview mode activated in this case, write:

   Please make an abstract of the contents. https://mountaineerbr.github.io/blog/2//

Also, adjust some constants in the script for it to work better in  some cases as these cases are becoming more normal as the script gets more robust and learns how to handle them properly.

Add some sanity newlines.

Fix unset retry/preview mode when there is a curl response error.

---