Avoid cluttering Bash shell history with dumps from text, pdf, and urls. Fix some preview mode issues.
Introduce two more envars for internal fun: $REPLY_CMD and $REPLY_CMD_VAR .
These will prevent polluting shell history with the dump or text files, urls and pdfs loaded when those are the last arguments of the user prompt (that is a recent feature).

Expand "~/" to "$HOME/" in user input in the function of the new feature described above.

Fix a bug in which previewing an URL dump would cause it to be re-dumped again

Also, fix stuff relating to keeping user edit of a text file, url or pdf dump when it is the last argument of the user prompt, *in the preview mode*! The bug was only when using the preview mode in this case.

Don't record dumped text files, urls and pdf in the bash shell history. Of course, do write them to the session history file!

Important fix in prev_tohistf() function when the session break was not yet recorded to the session history file!

Update _is_linkf(), delete redundant tests.

Fix some bugs when curl returned with error and the preview mode was on.

Add some more sanity new lines.

---