Add option -P (--print) as alias to -HH (print out history).
Typing -HH everytime for a nice print out of the current session is awkward, so
I alised -P to -HH and -PP to -HHH (to include commented out entries of the session).

Update man and help docs, and auto completion scripts.

Set option -P (and -HH) and chat command "/print" to print with markdown while in chat mode if markdown option is set!

Improve the set_mdcmdf() to set the markdown software to be used, just really set the name so that it can be shown to the user, if none set manually or does not exist on the system...

---