Add Whisper temperature parameter in the configuration file.
The Whisper function would take the temperature from the chat and cmpls models. We default it to 0 unless it is set from the command line option or from the config file with the name "OPTTW".

This "bug" was only exposed when Whisper was used as a stand-alone funtion (commadn line options -wwWW) and the temperature was set in a config file!

This should improve how Whisper functions integrate with the chat mode and when working alone.

If the sole argument to Whisper is either `retry`, `last` or `.`, retry whisper request with the last recorded audio.

---