Improve User eXperience with the new text/pdf/url dump feature loop.
Improve the new text/url/pdf dump feature.

Improve getting the exit code `from cmd_runf()` "shell command" in the new text/url/pdf dump feature to be able to run the
appropriate code depending on the user feedback.

Recognise when a prompt will trigger the new cat text/pdf/url feature and
skip the main confirmation prompt which is unecessary the dump feature
will be asked if the user wants to edit the dump.

Improve the UX when dealing with the new text/pdf/url dump feature loops,
specially when user wants to redo the prompt.

Adjust code to fit the new text/pdf/url dump feature in the "preview mode" better.

---