Exec demo mode when no command lie option is given.
This basically sets the plain text options and the option to exit on the first run. This is probably a demo mode, but I need to test it further.

Anyhow, the demo mode uses the plain text completion interface.

I need  do some more checking, specially with the OPTRESUME variable but for now it is fine.

This is a good basic presentation of what plain text completion endpoint can do!

---