Add Anthropic (Claude models), new option to set a seed from the command line, bug fixes and improvements.
Add Anthropic integration.

Add seed and image style settings.

Option `--seed [integer]`, same as `-j [integer]`, and positional parameter option to set image generation to vivid instead of natural, Dalle-3.

Add Anthropic, Mistral and Groq model globs  to the function which sets up maximum model tokens, and also, add known vision models to list.

Some endpoints do not yet offer vision capabilities, such as Mistral and Groq. Other integrations are ready to work with images!

Improved interface of insert (FIM, fim in the middle) mode.

Add option to generate an image as vivid or natural, as positional argument to option -i.

Fix some general bugs in the loop interface.

Set a higher priority for insert mode (options -q and -qq), this is higher than checking the model vision features in the multi turn loop.

Improve Groq and Mistral integrations a little more.

Update main jq processing syntax to process the response block.
Hopefully this update will avoid print even more types of errors, and also,
I added more finish reasons we dont want to be printed.

Add long option --version.

Add fix when resuming a session. Sometimes, the break session flag may have been written without actually user and assistant turns afterwards, meaning, that the script would not work correctly if option --continue -C was set.

This may happen when the script is killed or interrupted, but mostly it should not print a session break until there is an response to be recorded!

So we print a alert message to say that a session break was removed from the history file.

The script should be working very well, I tested all endpoints this afternoon
and adjusted some outdated code.
I will try not to do any big changes anymore, no more code refactoring!
If the script becomes broken, we either revert or archive the project.

Below is a  table with the results of the testing I did this afternoon.

The first OK means the user interface test and response was printed successfully to the screen, meaning the response json was processed correctly.

OK-OK means I was able to test when the token count from response json matched the sum of the recorded tokens in the history file, which is quite important. All OK.

The symbol "-" means that some integration services do not offer those endpoints, and "?" means I did not test (but the vision feature should be working OK).

---