Update documentation, add option --version to it.
Test curl for --fail-with-body, which is great for debugging.

Option -Z now can be set -ZZ or -ZZZ to print out the last response, the last to older response, and google full response (raw JSON).

Update some spinner characters to us ein different situations.... May revert if I find that is over the top as to animations! rs

Re add option of --fail-with-body for curl, do a test to check if curl supports --fail-with-body, otherwise, use --fail.

---