Optimise bash globs in tests, using [[ -z and [[ -n.
We were using very sophisticated extended globs do perform some tests tosee if input was empty, and that is not needed, really. So this makes the script a little more strict in the sense that beware of white spaces in input because they will be caught in the tests now. This improves bash speed a lot in some cases!

Also make adding a prompt as SYSTEM or as USER with the notation ":PROMPT" and "::PROMPT", respectively, must not contain white spces before the colon : or doble colon ::.

---