Fix help-assit function to only activate on stricter conditions (see chat command help)
It was catching /home/usr/file.pdf for example, which is just bad. So, now "/h somoething" just will not catch anything as a command and proceed to prompt submission, unless that is a file type which will be dumped, in which case that will be recognised as such and correctly.

---