Improve the chat mode with voice input and tts output automatic modes -ccwzvv and -wzvv.
Fix command /replay to replay the defaults tts output file if no file in the replay array.

Add experimental code to try and detect when the user stops talking in voice input mode. This works for the whisper function. So that we try to detect a pause (a silent period) after the user stops talking when `options -vv` are set.

In this way, it is closer to the hands-free experience an Andorid user may have with the conversation mode of the ChatGPT app.

This is experimental.

Known Bugs:

* Noise in background prevent automatic silence detection
* If the user does not starts talking as soon as the recorder starts audio capture, the silence detection mechanism will interrupt the recording
* If user talks too slowly, it risks the recording interruption prematurely

Some of the above bugs are reasonable while other may be worked on.

In any case, the user can press any key to end the recording, as usual.

On other fronts, `option -z` (TTS) combined with `option -v` now defaults to playing the response from TTS API as soon as there is some buffer to be played, unless Termux environment is detected as this is not recommended for Termux users because in my tests, the players cannot work properly with the file that is being live-appended.

Now, chat mode with `options -wv` (when _not combined_ with `option -z`⁾, will print to stderr the countdown of the automatic pause for the user to be able to read the response text before automatically progressing to whisper mode.

Fix a bug in termux: dont automatically play the receiving audio stream (TTS), because in my tests in two different Termux system, it always quits playing the response audio in the middle, so we better wait until we have got the full response.

Also, the silence detection mechanism is disabled because it cannot work realiably under Termux. Termux recorder does not add a moov atom to the input-writing file, so we cannot use neither ffmepg or sox to detect silence in the record as they error out when reading the appending file.

Tentatively add a bellf() function to play a bell sound when it is the user turn to speak in record mode. This is under tests, we dont want to overload the system with user warnings.

After taking a photo with termux wrapper in Android, resize and rotate image if necessary with image magick so that the image fits OpenAI image size guidelines. Note that this feature is only used here because while useful to speed up the uploading process with image sthat are not bigger than needed, some other API services do support larger images, such as Gemini, for as far as I can tell for what I read in some API docs.

---