On Termux, disown pulseaudio daemon forked shell. Avoid printing multimodal "invalid input" errors in some cases.
When the script tries to set module-sles-source in Termux, try to keep it alive when the script exits or is killed by disowning pulseaudio daemon.

Avoid printing multimodal input errors when we detect there is spaces in input.Input in this case must be a command as text in the user prompt rather than multimodal input. Anyways, when an input is accepted, there is user feedback.

---