Add support to easily dump text from .doc, .xls, etc file types.
Important: fix the request block json that affected a few specific situations when using the Anthropic API and its integrational code.

Use LibreOffice when it is availiable to convert lots of different file types to text, such as word documents and spreadsheets.

Improved how the automatic commands for pdf and doc dumps disable themselves in case no pdf-to-text or libreoffice suites are not found in the ssystem. This avoids the overhead of checking for the existence of the required software beforehand.

Add some comments in the source code with stuff that I am currently studying and peerhaps implement them in the code in the future, if needed.

Add sanity newlines.

Fix detecting the chat "/cat command" when there is arguents to it (optimisation regression).

Minor adjustments of some varible substitution indexes and constants.

Much improved handling of files with spaces in paths for all the text dump functions (/cat, /pdf, and /doc).

New is_docf() function, and upgraded is_txturl() test function.

```
function _is_docf
{
	typeset -l ext=$1
	case "$ext" in
	*.doc|*.docx|*.docm|*.dot|*.dotx|*.dotm|*.odt|*.ott|*.oth|*.rtf|*.uot|*.uof|
	*.xls|*.xlsx|*.xlsm|*.xlt|*.xltx|*.xltm|*.ods|*.ots|*.fods|*.uos|
	*.ppt|*.pptx|*.pptm|*.pot|*.potx|*.potm|*.pps|*.ppsx|*.ppsm|*.odp|*.otp|
	*.dif|*.slk|*.dbf|*.odg|*.otg|*.sxd|*.std|*.odf) :;;
	*) false;;
	esac;
}

```

---