Vision / multimodal models change of input handling. Silently process filenames appended to the end of user prompt.
Before, it was required to separate filenames with a pipe "|" when a filename had white spaces.
Now, pipe _is not_ used anymore as an operator-separator to delimit multimodal input.

However, filenames with white spaces are now required to have backslash-escaped white spaces in the middle of the filepath!

This is the way shell scripts must work as it resonates with other shell-way of doing things.

Improve how chat commands are executed when they invoke the "/shell" command internally.

Turning on `$PSKIP` is very powerful in how it changes the procedure to take in the chat turn.
It must be strictly controlled, that is why now we have many added many safety measures to
unset `$PSKIP` after it it serves it purpose.

This was not so much of an evident problem before but now that we execute the "/shell" command
a lot for automatic features (text dumps, mostly), we must take heed of `$PSKIP` closely.

This improves general user interface for redo / edit functions of the current user prompt.

Improve the text editor interface a little, specially when user selects prompt "redo".

Improve on the last few code updates.

Improve the "/pick" to allow for picking filenames with white spaces.

---