Update is_mdf() to be more strict when checking for markdown in response.
As this function is used to auto-detect markdown in response and enable markdown rendering, it turns out that some answers only have ordered or unordered lists, and if that is the *only* markdown, it is not worth it turning on markdown automatically. However, is the lists are idented, then markdown will be turned on.

Maybe in the future we will remove the markdown lists regex, because really just turn on markdown rendering when tehre is enough markdown to render!

---