Comment out executing trim_silencef() function to remove silence from voice record.
The threshold is probably too low to actually remove any significant amount of "silence" from the voice record.

This works well with professionally recorded audio to remove silence periods, but voice record in random environments usually have a lot of background noise.

So I am commenting this function out as it is probably an overhead at this time. I want to test it more to see if it actually makes any significant "silence" removal.

---