Bug fixes + update default models for Groq and Google Gemini services.
Update default models

GROQ: llama-3.1-70b-versatile

GEMINI: gemini-1.5-pro-latest

---

Code cleaning. We clean some old code which must not affect how it works.
Mostly minor cleaning changes.

---

Fix very important bug that would really only show up when running the script with options "-vv",
in _sysmsgf() (old __sysmsgf) because it would return with 1 in case $OPTV was not less than 2.

This should be nothing but we use "! _sysmsgf" sometimes to return from other functions with
an error code, but it would just flop the returning value to truw when it was false (when
options -vv were set).

---

The function naming convention of the script was a little awkward and
many function names has double underscores before the function name.
We reduced the use of double underscores in name function to make it
more readble.

The confusion derives from what is now the _sysmsgf() function which was one of the first
functions to be written and was kind of a base for the following function namings.

Also, chat mode would increase +1 $OPTV value, and _sysmsgf() heeds it,
but automatically increasing $OPTV in chat mode proved to be a bad idea
and was reverted some time there after.

However, the logic underlying _sysmsgf() was flawed at the beggining and it has changed
so it has been time to change all functions that have weird names.

Basically, one underscore means it is a helper function *or* the raw function
of a feature, and two underscores means a helper of a helper function.

Not perfect logic but hopefully it is now aligned.

---

The The `+=' assignment operator is now used less frequently in the code to append strings to strings.
We only use this operator to append items to arrays and in arithmetic ops.

Why?

As this bash feature is so little used, and there is a more inter shell compatible ways of performing the same operations, and
albeit _that is not_ the goal to be strictly bourne-shell compatible, we prefer to decrease the use of bashisms as much as possible.

bash-3.1-alpha1:

i.  The `+=' assignment operator (append to the value of a string or array) is
    now supported for assignment statements and arguments to builtin commands
    that accept assignment statements.

---