Bug fixes in the "dump text/pdf/doc/url" feature and in the text editor interface.
The somewhat recent code for "dump text/pdf/doc/url" feature needed some adjustments
to work well when the dump returned empty OR with an exit code.

It was necessary to check for the error code outside the scope of exit codes
of the user confimation prompt options, and outside of the range of
exit codes  our script itself uses for internal signaling.

The script internal exit codes range from 180 to 220, or a little more
narrowly, between 190 to 210.

Improve the text editor interface in how the "-xx" chat command or
pressing "x" in user prompts would
return to the text editor with the correct INPUT.

Add new OpenAI model chatgpt-4o to known list of models, so that model capacity
has the right setting.

Update known model pricing.

Renamed the configuration variable $COST_CUSTOM to $MOD_PRICE, as teh previous name for this variable is really bad...

Added the shortcut to exit the script with pressing "Q" (uppercase) in the
user confirmation prompts (only when option -v is not set).

---