Fix Termux auto-set pulseaudio "module-sles-source".  Defaults to NO.
We revert defaulting to NO, do not set pa "sles source" automatically.

We keep the warnings as this is probably a setting the Termux user wants to add to his pulseaudio configuration.

Only print the warnings when options -vv are set, so we dont disrupt the old Termux way ofrecording andplaying which works just fine...

The pulseaudio with sles module is only required for the silence-detection feature, so lets not set it unless needed!

---