Bug fixes and improvements.
Fix the function session_globf() when the argument is "current", and "current" points to a custom filepath for the chat history file.

Remove trailing spaces from the argument to session_mainf().

Fix setting INSTRUCTIONS in a new chat when using chat command "/break". Instructions were not being backed up correctly.

---