Edit bash read input in text editor: fix regression introduced on 1 Sep 2024 on commit d146f212fb1990030da4f6a540c3757069a19d9e (GitLab).
This commit addressed some issues when a command output was to be edited in the text editor interface, however adjusting the user confirmation prompt in the bash read function was not fixed to properly follow the new rules.

This is now fixed to be able to edit current bash read input in the text editor interface, as expected!

---