Fix important old bug that uneedlessly escaped system and user prompts in the append function.
This bug has been around since:

17 February 2024 - GitLab - 6af160c157b36a117589b26fb07cbb60ea331823

When we were implementing Google AI integration and also changing how we prefer to record results to the history file.

This bug was pretty low key and hardly affected any other part of the script other than appending system/user messages to the message stack manually, and those messages were usually one liners or few-liners in my own testing.

More recently, this bug also affected the "/sh:", "/url:", and "/pdf:" commands noticeably, which output would be unnecessarily escaped twice, and thus this bug is now fixed!

The latter task at the time introduced lots of bugs because the script up til then, used the record file from the moment it was started up, so it  relied on the record file to build a request block even before the first request was made. This all changed in the months after.

Now, the script is mostly free from the bugs introduced in the way we record to the history file since recently. This is  a late bug catch but it is very important!

---