Revert defaults model to "gpt-4o", add "print" option when searching for a session.
When searching for a session in a history file, only the tail snippet of a session is shown. However, now, pressing "[p]" will print out the the session to stderr so the user can make sure that is teh right session before actually commiing to copy it to the end of the history file for resuming it.

This should avoid copying the session to the tail of the history file when all the user wants is to check if that is the right session!

---