Be a little more restrcitive when invoking end-of-prompt commands.
This applies to `/pick`, for example. So now, as this may be a start-of-prompt comamnd as well, we check for an `$IFS` char before the command to make sure this may really be a end-of-prompt command.

This limits the `/pick` file picker invokation a little but we dont see a problem there.

---