Add support for gtp-4o-audio-preview models. Various improvements and optimisations. Bug fixes.
Add support for gtp-4o-audio-preview. When calling the script, setting option '-w' will record user voice, otherwise he can set an audio file as input manually at the end of the prompt which should be loaded with the request.

To receive audio as output, needs to set option '-z' at the command line.

Fix some problems with the chat command "/h[regex]". It is now more strict and needs to have a space between the command and the argument, such as "/h [regex]". This avoids some problems with filename with full path "/home/...".

Optimise some script internal parameters for it to be more responsive to user interaction.

Much optimised incoming stream processing (json) chunks! Makes the response fly on the screen now! I was using pure bash built-ins for substring manipulation at a while loop,but sed proves to be the perfect solution!

Fix issue with image file that had white spaces in them and were added to the propmt with chat command "/img [file.jpg]". If the file had white spaces in the name, it would not record to the history file correctly anymore, since we redesigned how the script handles image input with spaces in the filepath..

However, the mentioned change in the script should have been backwards compatible with the old filepath with spaces recording system, so I fixed it! It is now possible to append images and audio in the old-fashioned way as well `My text prompt. | path/to the file.jpg`!

Fix various bugs and much improved user interface flow and error handling, if I say so myself!

The script recovered some of the old and witty ideias of how the user interface ought to work..

Update some code for better readability and future maintenance efforts.

Also, prime the code of the script to accept multiple modalities  of input in the future (text, audio, and image).

Update help and man pages.

As the preview audio-model has only the option to stream back audio in PCM16 format, I am still a little unsure how to handle this type of file the most proper way possible. So for now, we do record the pcm16 stream as "[audio_filename].pcm" even though it might probably be best to set the file name as ".raw"...?

---

KNOWN OR PRESUMED BUGS: token count of the audio-models kind of works but not perfectly. It is also hard to try and decide what to record, because audio models have reasoning tokens and audio tokens, and audio tokens may be present for the user input and/or as part of the output!

---

TO-DO for next releases:
  #add test for filename with white spaces separated by pipe in test suite.
  #consider checking simply $MULTIMODAL>0, instead of $MULTIMODAL>1.
  #reconsider the additional audio-model related defaults $INSTRUCTION_CHAT text appended when all configs are defaults. Simplify it?

---